Addressed comments, add background colors and organized the data into separate files for better structure.
