+++ data/conference.js
@@ -1,0 +1,61 @@
+export const CONFERENCE = {
+  title: 'PyConf Hyderabad 2025',
+  description:
+    'PyConf Hyderabad 2025 is the regional gathering of Pythonistas and the largest Python Conference in Hyderabad.',
+  announcement:
+    'Hyderabad’s biggest Python conference is on the way! Get ready for an unforgettable experience and stay tuned for updates!',
+  copyrightTitle: 'Copyright @ 2025 PyConf Hyderabad',
+  dates: '22 Feb - 23 Feb',
+};
+
+export const ASSETS = {
+  navbarLogoUrl: '/images/navbarLogo.svg',
+  navbarLogoAlt: 'PyConf Hyderabad 2025 Python Logo',
+  logoUrl: '/images/logo.svg',
+  logoAlt: 'PyConf Hyderabad 2025 Logo',
+  monumentsSketchUrl: '/images/monumentsSketch.svg',
+  monumentsSketchAlt:
+    'Illustration showcasing various iconic monuments of Hyderabad.',
+};
+
+export const KEY_LINKS = {
+  volunteerFormLabel: 'BECOME A VOLUNTEER',
+  volunteerFormUrl: 'https://forms.gle/aceBohpaYLWef4eU9',
+  cfpLabel: 'SUBMIT YOUR PROPOSAL',
+  cfpUrl: 'https://www.papercall.io/pyconfhyd2025',
+  interestedInSponsoringLabel: 'INTERESTED IN SPONSORING',
+  interestedInSponsoringUrl: 'https://forms.gle/d5cg5WpayP9mEfAm9',
+};
+
+export const SOCIALS = [
+  {
+    name: 'Envelope',
+    url: 'mailto:contact@hydpy.org',
+    ariaLabel: "Hyperlink to HydPy's email.",
+  },
+  {
+    name: 'LinkedIn',
+    url: 'https://www.linkedin.com/company/pyconf-hyderabad',
+    ariaLabel: "Hyperlink to PyConf Hyderabad's LinkedIn page.",
+  },
+  {
+    name: 'GitHub',
+    url: 'https://github.com/hydpy',
+    ariaLabel: "Hyperlink to HydPy's GitHub page.",
+  },
+  {
+    name: 'X',
+    url: 'https://x.com/pyconfhyd',
+    ariaLabel: "Hyperlink to PyConf Hyderabad's X page.",
+  },
+  {
+    name: 'Instagram',
+    url: 'https://www.instagram.com/pyconfhyd',
+    ariaLabel: "Hyperlink to PyConf Hyderabad's Instagram page.",
+  },
+  {
+    name: 'Discord',
+    url: 'https://discord.com/invite/C3pnRHMyWJ',
+    ariaLabel: "Hyperlink to PyConf Hyderabad's Discord server.",
+  },
+];
