--- data/details.js
@@ -1,118 +1,0 @@
-export const CONFERENCE = {
-  title: 'PyConf Hyderabad 2025',
-  description:
-    'PyConf Hyderabad 2025 is the regional gathering of Pythonistas and the largest Python Conference in Hyderabad.',
-  announcement:
-    'Hyderabad’s biggest Python conference is on the way! Get ready for an unforgettable experience and stay tuned for updates!',
-  copyrightTitle: 'Copyright @ 2025 PyConf Hyderabad',
-  dates: '22 Feb - 23 Feb',
-};
-
-export const ASSETS = {
-  navbarLogoUrl: '/images/navbarLogo.svg',
-  navbarLogoAlt: 'PyConf Hyderabad 2025 Python Logo',
-  logoUrl: '/images/logo.svg',
-  logoAlt: 'PyConf Hyderabad 2025 Logo',
-  monumentsSketchUrl: '/images/monumentsSketch.svg',
-  monumentsSketchAlt:
-    'Illustration showcasing various iconic monuments of Hyderabad.',
-};
-
-export const ABOUT_HYDPY = {
-  title: 'Meet our Organizers: HydPy',
-  description:
-    'Hyderabad Python User Group (HydPy) is community driven group to promote Python in Hyderabad. HydPy has been conducting monthly Meetups & workshops in Hyderabad for last 9+ years and it has over 15000 members on meetup. We work together with other communities in particular, python and in general Open Source Software (OSS) related groups within Hyderabad and the rest of India to further expand the usage and increase users of OSS. HydPy is a recognised regional community by global Python Software Foundation (PSF).',
-  hydPyLogoUrl: '/images/hydPyLogo.svg',
-  hydPyLogoAlt: 'HydPy Logo',
-  hydPyUrl: 'https://www.hydpy.org/',
-  hydPyLinkText: 'Dive Into HydPy',
-};
-
-export const KEY_LINKS = {
-  volunteerFormUrl: 'https://forms.gle/aceBohpaYLWef4eU9',
-  cfpUrl: 'https://www.papercall.io/pyconfhyd2025',
-  interestedInSponsoringUrl: 'https://forms.gle/d5cg5WpayP9mEfAm9',
-};
-
-export const NAV_ITEMS = [
-  {
-    name: 'Home',
-    path: '/',
-    target: '_self',
-  },
-  {
-    name: 'Code Of Conduct',
-    path: 'https://docs.google.com/document/d/1yjPaVdsQCbowelwI1LlGco3CFGG3vZKhGiMYXDOOKrY',
-    target: '_blank',
-  },
-];
-
-export const PYCON_INDIA_HYDPY_PAST_EDITIONS = {
-  sectionTitle: 'PyCon India Editions: A Collective Effort Led by HydPy',
-  editions: [
-    {
-      year: 2023,
-      url: 'https://in.pycon.org/2023/',
-      ariaLabel: 'PyCon India 2023 website',
-    },
-  ],
-};
-
-export const PAST_EDITIONS = {
-  sectionTitle: 'Past Editions',
-  editions: [
-    {
-      year: 2022,
-      url: 'https://pyconf.hydpy.org/2022',
-      ariaLabel: 'PyConf Hyderabad 2022 website',
-    },
-    {
-      year: 2020,
-      url: 'https://pyconf.hydpy.org/2020',
-      ariaLabel: 'PyConf Hyderabad 2020 website',
-    },
-    {
-      year: 2019,
-      url: 'https://pyconf.hydpy.org/2019',
-      ariaLabel: 'PyConf Hyderabad 2019 website',
-    },
-    {
-      year: 2017,
-      url: 'https://pyconf.hydpy.org/2017',
-      ariaLabel: 'PyConf Hyderabad 2017 website',
-    },
-  ],
-};
-
-export const SOCIALS = [
-  {
-    name: 'Envelope',
-    url: 'mailto:contact@hydpy.org',
-    ariaLabel: "Hyperlink to HydPy's email.",
-  },
-  {
-    name: 'LinkedIn',
-    url: 'https://www.linkedin.com/company/pyconf-hyderabad',
-    ariaLabel: "Hyperlink to PyConf Hyderabad's LinkedIn page.",
-  },
-  {
-    name: 'GitHub',
-    url: 'https://github.com/hydpy',
-    ariaLabel: "Hyperlink to HydPy's GitHub page.",
-  },
-  {
-    name: 'X',
-    url: 'https://x.com/pyconfhyd',
-    ariaLabel: "Hyperlink to PyConf Hyderabad's X page.",
-  },
-  {
-    name: 'Instagram',
-    url: 'https://www.instagram.com/pyconfhyd',
-    ariaLabel: "Hyperlink to PyConf Hyderabad's Instagram page.",
-  },
-  {
-    name: 'Discord',
-    url: 'https://discord.com/invite/C3pnRHMyWJ',
-    ariaLabel: "Hyperlink to PyConf Hyderabad's Discord server.",
-  },
-];
