+++ data/hydPy.js
@@ -1,0 +1,9 @@
+export const ABOUT_HYDPY = {
+  title: 'Meet our Organizers: HydPy',
+  description:
+    'Hyderabad Python User Group (HydPy) is community driven group to promote Python in Hyderabad. HydPy has been conducting monthly Meetups & workshops in Hyderabad for last 9+ years and it has over 15000 members on meetup. We work together with other communities in particular, python and in general Open Source Software (OSS) related groups within Hyderabad and the rest of India to further expand the usage and increase users of OSS. HydPy is a recognised regional community by global Python Software Foundation (PSF).',
+  hydPyLogoUrl: '/images/hydPyLogo.svg',
+  hydPyLogoAlt: 'HydPy Logo',
+  hydPyUrl: 'https://www.hydpy.org/',
+  hydPyLinkText: 'Dive Into HydPy',
+};
