+++ data/navItems.js
@@ -1,0 +1,12 @@
+export const NAV_ITEMS = [
+  {
+    name: 'Home',
+    path: '/',
+    target: '_self',
+  },
+  {
+    name: 'Code Of Conduct',
+    path: 'https://docs.google.com/document/d/1yjPaVdsQCbowelwI1LlGco3CFGG3vZKhGiMYXDOOKrY',
+    target: '_blank',
+  },
+];
