+++ data/pastEditions.js
@@ -1,0 +1,36 @@
+export const PYCON_INDIA_HYDPY_PAST_EDITIONS = {
+  sectionTitle: 'PyCon India Editions: A Collective Effort Led by HydPy',
+  editions: [
+    {
+      year: 2023,
+      url: 'https://in.pycon.org/2023/',
+      ariaLabel: 'PyCon India 2023 website',
+    },
+  ],
+};
+
+export const PAST_EDITIONS = {
+  sectionTitle: 'Past Editions',
+  editions: [
+    {
+      year: 2022,
+      url: 'https://pyconf.hydpy.org/2022',
+      ariaLabel: 'PyConf Hyderabad 2022 website',
+    },
+    {
+      year: 2020,
+      url: 'https://pyconf.hydpy.org/2020',
+      ariaLabel: 'PyConf Hyderabad 2020 website',
+    },
+    {
+      year: 2019,
+      url: 'https://pyconf.hydpy.org/2019',
+      ariaLabel: 'PyConf Hyderabad 2019 website',
+    },
+    {
+      year: 2017,
+      url: 'https://pyconf.hydpy.org/2017',
+      ariaLabel: 'PyConf Hyderabad 2017 website',
+    },
+  ],
+};
