+++ data/sponsors.js
@@ -1,42 +1,16 @@
+export const SPONSORS_INFO = {
+  title: 'Sponsors',
+  description:
+    'PyConf Hyderabad is completely driven by volunteers. Sponsoring the event helps to sustain and grow the conference as well as the Community. Sponsor help in making the conference affordable, and maintaining the inventory for the conference.',
+};
+
 export const SPONSORS = {
-  platinumSponsors: [
+  Platinum: [
     {
       id: 1,
       name: 'Python Software Foundation',
       description: 'The organization behind Python programming language',
-      website: 'https://www.python.org/psf/',
-      logoUrl: '/images/logo.svg',
-    },
-  ],
-  goldSponsors: [
-    {
-      id: 2,
-      name: 'Django Software Foundation',
-      description: 'Supporting and maintaining Django web framework',
-      website: 'https://www.djangoproject.com/foundation/',
-      logoUrl: '/images/logo.svg',
-    },
-    {
-      id: 3,
-      name: 'JetBrains',
-      description: 'Creator of PyCharm and other development tools',
-      website: 'https://www.jetbrains.com/',
-      logoUrl: '/images/logo.svg',
-    },
-  ],
-  silverSponsors: [
-    {
-      id: 4,
-      name: 'FastAPI',
-      description: 'Modern, fast web framework for building APIs with Python',
-      website: 'https://fastapi.tiangolo.com/',
-      logoUrl: '/images/logo.svg',
-    },
-    {
-      id: 5,
-      name: 'Streamlit',
-      description: 'The fastest way to build data apps in Python',
-      website: 'https://streamlit.io/',
+      hyperLink: 'https://www.python.org/psf/',
       logoUrl: '/images/logo.svg',
     },
   ],
