+++ src/app/layout.js
@@ -1,5 +1,5 @@
 import './globals.css';
-import { CONFERENCE, ASSETS } from '@/details';
+import { CONFERENCE, ASSETS } from '@/conference';
 import Header from '@/components/Header';
 import Footer from '@/components/Footer';
 import { ThemeToggle } from '@/components/ThemeToggle';
@@ -20,7 +20,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className="flex flex-col min-h-screen bg-gray-50 text-gray-900 dark:bg-gray-800 dark:text-gray-100">
+      <body className="flex flex-col min-h-screen bg-background-light dark:bg-background-dark text-gray-900 dark:text-gray-100">
         <Header themeToggle={<ThemeToggle />} />
         <main className="flex-grow">{children}</main>
         <Footer />
