+++ src/components/MeetOrganizers.jsx
@@ -1,19 +1,19 @@
-import { ABOUT_HYDPY } from '@/details';
 import Image from 'next/image';
 import Link from 'next/link';
 
 import { Heading, Paragraph, Span } from '@/components/Typography';
+import { ABOUT_HYDPY } from '@/hydPy';
 
 export default function MeetOrganizers() {
   return (
-    <div className="flex flex-col items-center py-6 w-11/12 lg:w-5/6 mx-auto">
+    <section className="flex flex-col items-center py-6 w-11/12 lg:w-5/6 mx-auto">
       <Heading
-        level={2}
+        level={1}
         className="text-center my-8 text-secondary-600 dark:text-secondary-400"
       >
         {ABOUT_HYDPY.title}
       </Heading>
-      <div className="flex flex-col gap-3">
+      <div className="flex flex-col gap-3 items-center">
         <div className="flex flex-col items-center gap-3">
           <Image
             className="rounded-lg"
@@ -30,10 +30,10 @@ export default function MeetOrganizers() {
             <Span>{ABOUT_HYDPY.hydPyLinkText}</Span>
           </Link>
         </div>
-        <Paragraph className="text-gray-600 dark:text-gray-400">
+        <Paragraph className="text-gray-600 dark:text-gray-400 lg:w-11/12 text-center">
           {ABOUT_HYDPY.description}
         </Paragraph>
       </div>
-    </div>
+    </section>
   );
 }
